TestShop/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
openpyxl==3.1.2
PyJWT==1.7.1
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
